fluid poster - transformation array fixed
